Create a base class for Api Endpoints and add some docblocks
